+++ Desafio1/emprendedor3.py
@@ -1,7 +1,7 @@
 print("Para calcular las utilidades necesitaremos que nos proporciones algunos datos.")
 
 precioSuscripcion = float(input("Ingresa el valor de las suscripcion:\n"))
-numeroUsuarios = float(input("Ingresa el número de usuarios:\n"))
+numeroUsuarios = int(input("Ingresa el número de usuarios:\n"))
 gastosTotales = float(input("Ingresa el valor de los gastos totales:\n"))
 utilidadesAnterior = float(input("Ingrese las utilidades del año anterior:\n"))
